03. Next Components Server and Client | 16. Some More Refactoring
