+++ app/src/app/layout.tsx
@@ -5,8 +5,8 @@ import './globals.css';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'OpenTable',
+  description: 'OpenTable Restaurant',
 };
 
 export default function RootLayout({
