+++ app/src/app/reserve/[slug]/components/ReserveHeader.tsx
@@ -1,0 +1,22 @@
+export default function ReserveHeader() {
+  return (
+    <div>
+      <h3 className="font-bold">You're almost done!</h3>
+      <div className="mt-5 flex">
+        <img
+          src="https://images.otstatic.com/prod1/49153814/2/medium.jpg"
+          alt=""
+          className="w-32 h-18 rounded"
+        />
+        <div className="ml-4">
+          <h1 className="text-3xl font-bold">Aiāna Restaurant Collective</h1>
+          <div className="flex mt-3">
+            <p className="mr-6">Tues, 22, 2023</p>
+            <p className="mr-6">7:30 PM</p>
+            <p className="mr-6">3 people</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
