+++ app/src/app/restaurant/[slug]/components/RestaurantDescription.tsx
@@ -1,0 +1,13 @@
+export default function RestaurantDescription() {
+  return (
+    <div className="mt-4">
+      <p className="text-lg font-light">
+        The classics you love prepared with a perfect twist, all served up in an
+        atmosphere that feels just right. That’s the Milestones promise. So,
+        whether you’re celebrating a milestone, making the most of Happy Hour or
+        enjoying brunch with friends, you can be sure that every Milestones
+        experience is a simple and perfectly memorable one.
+      </p>
+    </div>
+  );
+}
