+++ app/src/app/restaurant/[slug]/components/RestaurantHeader.tsx
@@ -1,0 +1,11 @@
+export default function RestaurantHeader() {
+  return (
+    <div className="h-96 overflow-hidden">
+      <div className="bg-center bg-gradient-to-r from-[#0f1f47] to-[#5f6984] h-full flex justify-center items-center">
+        <h1 className="text-7xl text-white captitalize text-shadow text-center">
+          Milestones Grill (Toronto)
+        </h1>
+      </div>
+    </div>
+  );
+}
