+++ app/src/app/restaurant/[slug]/components/RestaurantRating.tsx
@@ -1,0 +1,13 @@
+export default function RestaurantRating() {
+  return (
+    <div className="flex items-end">
+      <div className="ratings mt-2 flex items-center">
+        <p>*****</p>
+        <p className="text-reg ml-3">4.9</p>
+      </div>
+      <div>
+        <p className="text-reg ml-4">600 Reviews</p>
+      </div>
+    </div>
+  );
+}
