+++ app/src/app/restaurant/[slug]/components/RestaurantTitle.tsx
@@ -1,0 +1,7 @@
+export default function RestaurantTitle() {
+  return (
+    <div className="mt-4 border-b pb-6">
+      <h1 className="font-bold text-6xl">Milesstone Grill</h1>
+    </div>
+  );
+}
